[TEST] restructure tests
test for functions from a specific python file are grouped together in the same structure under unit tests, larger workflow tests are add in e2e testing
